:recycle: モデルをパブリックなものとそうでないものに分離
